requirements.txt
numpy==1.17.0
pandas
pyyaml
prettytable==0.7.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0